fix behat suite

+++ features/csrf_protection.feature
@@ -7,7 +7,7 @@ Feature: CSRF protection for unsafe requests
         Given I write in "App/Controller/FooController.php":
         """
         <?php namespace App\Controller {
-            class FooController extends \Knp\RadBundle\Controller\Controller
+            class FooController
             {
                 public function showAction() {}
 
@@ -23,16 +23,18 @@ Feature: CSRF protection for unsafe requests
         App:Foo:show: ~
         App:Foo:delete:
             defaults: { _check_csrf: true }
+            requirements: { method: delete }
         App:Foo:deleteUnsafe:
             defaults: { _check_csrf: false, _controller: App:Foo:delete }
+            requirements: { method: delete }
         """
         And I write in "App/Resources/views/Foo/show.html.twig":
         """
         {% extends 'KnpRadBundle:Layout:h5bp.html.twig' %}
         {% block body %}
-            <a {{ link_attr('delete') }} href="{{ path('app_foo_delete') }}">Delete</a>
-            <a data-method="delete" data-csrf-token="test" href="{{ path('app_foo_delete') }}">Invalid Delete</a>
-            <a data-method="delete" href="{{ path('app_foo_deleteUnsafe') }}">Unsafe Delete</a>
+            <a {{ link_attr('delete', false) }} href="{{ path('app_foo_delete') }}">Delete</a>
+            <a data-method="delete" data-no-confirm data-csrf-token="test" href="{{ path('app_foo_delete') }}">Invalid Delete</a>
+            <a data-method="delete" data-no-confirm href="{{ path('app_foo_deleteUnsafe') }}">Unsafe Delete</a>
         {% endblock body %}
         """
